fix(cmd): change file structure to cobra-cli mentioned convention
